+++ include/geometry.h
@@ -2,12 +2,12 @@
 #define GEOMETRY_H
 
 typedef struct{
-	double x, y, z;
-	double normSquared;
+	float x, y, z;
+	float normSquared;
 }Vector;
 
 typedef struct{
-	double x, y, z;
+	float x, y, z;
 }Point;
 
 typedef struct{
@@ -16,17 +16,17 @@ typedef struct{
 }Line;
 
 typedef struct{
-	double A, B, C, D;
+	float A, B, C, D;
 }Plane;
 
 
 // ───── POINT ─────
 
 // Constructor
-Point*  Point_init(double x, double y, double z);
+Point*  Point_init(float x, float y, float z);
 
 // Operations
-double  Point_distanceSquared(Point *a, Point *b);
+float  Point_distanceSquared(Point *a, Point *b);
 Point*  Point_translate(Point *p, Vector v);
 
 // Debug
@@ -37,17 +37,17 @@ int     Point_size(Point *p);
 // ───── VECTOR ─────
 
 // Constructor
-Vector  Vector_init(double x, double y, double z);
+Vector  Vector_init(float x, float y, float z);
 Vector  Vector_fromPoints(Point *a, Point *b);
 
 // Operations
 Vector  Vector_normalize(Vector v);
-Vector  Vector_scale(Vector v, double scalar);
+Vector  Vector_scale(Vector v, float scalar);
 Vector  Vector_sum(Vector v1, Vector v2);
 Vector  Vector_crossProduct(Vector a, Vector b);
-double  Vector_dot(Vector v1, Vector v2);
+float  Vector_dot(Vector v1, Vector v2);
 Vector  Vector_perpendicular(Vector v);
-Vector  Vector_rotate(Vector v, Vector axis, double angle);
+Vector  Vector_rotate(Vector v, Vector axis, float angle);
 int     Vector_equal(Vector v1, Vector v2);
 
 // Debug
@@ -61,7 +61,7 @@ int	  Vector_size(Vector v);
 Line*   Line_init(Point *origin, Vector direction);
 
 // Operations
-double  Line_Point_distance(Line *l, Point *p);
+float  Line_Point_distance(Line *l, Point *p);
 Point*  Line_projectionPoint(Line *l, Point *p);
 
 // Debug
@@ -71,7 +71,7 @@ void    Line_print(Line *l);
 // ───── PLANE ─────
 
 // Constructor
-Plane*  Plane_new(double A, double B, double C, double D);
+Plane*  Plane_new(float A, float B, float C, float D);
 
 // Operations
 Vector  Plane_vectorNormal(Plane *plane);
